ckks_engine::ckks
Struct CKKSDecryptor

pub struct CKKSDecryptor { /* private fields */ }

The CKKSDecryptor is used for decrypting ciphertexts generated in the CKKS encryption scheme. It supports decryption into integers, floating-point numbers, and strings. The decryption process involves reversing the encryption scheme using the secret key.

Implementations

impl CKKSDecryptor

pub fn new(sec_key: SecretKey, params: CkksParameters) -> Self

Creates a new CKKSDecryptor instance with the provided secret key and encryption parameters.

pub fn decrypt_as_int(&self, ciphertext: &Polynomial) -> Vec<i64>

Decrypts the ciphertext and returns the result as a vector of integers.

pub fn decrypt(&self, ciphertext: &Polynomial) -> Vec<f64>

Decrypts the ciphertext and returns the result as a vector of floating-point numbers.

pub fn decrypt_string(&self, ciphertext: &Polynomial) -> String

Decrypts the ciphertext and returns the result as a string.