ckks_engine::ckks
Struct CKKSEncryptor

pub struct CKKSEncryptor {
    pub pub_key: PublicKey,
    pub params: CkksParameters,
}

Fields
pub_key: PublicKeyparams: CkksParameters
Implementations

impl CKKSEncryptor

pub fn new(pub_key: PublicKey, params: CkksParameters) -> Self

pub fn encrypt_collection<T>(&self, plaintext: &[T]) -> Polynomial
where T: Into<f64> + Copy,

pub fn encrypt_value<T>(&self, plaintext: T) -> Polynomial
where T: Into<f64> + Copy,

pub fn encrypt_string(&self, plaintext: &str) -> Polynomial

impl CKKSEncryptor

pub fn homomorphic_add( &self, cipher1: &Polynomial, cipher2: &Polynomial, ) -> Polynomial

pub fn homomorphic_subtract( &self, cipher1: &Polynomial, cipher2: &Polynomial, ) -> Polynomial

pub fn homomorphic_multiply( &self, cipher1: &Polynomial, cipher2: &Polynomial, ) -> Polynomial

pub fn homomorphic_negation(&self, cipher1: &Polynomial) -> Polynomial

pub fn homomorphic_ceil(&self, cipher: &Polynomial) -> Polynomial

pub fn homomorphic_floor(&self, cipher: &Polynomial) -> Polynomial

pub fn homomorphic_round(&self, cipher: &Polynomial) -> Polynomial

pub fn homomorphic_truncate(&self, cipher: &Polynomial) -> Polynomial

pub fn homomorphic_reciprocal( &self, cipher: &Polynomial, iterations: u32, ) -> Polynomial

pub fn homomorphic_divide( &self, cipher1: &Polynomial, cipher2: &Polynomial, ) -> Polynomial

pub fn homomorphic_exponentiation( &self, cipher: &Polynomial, exponent: u32, ) -> Polynomial

impl CKKSEncryptor

pub fn homomorphic_length(&self, encrypted_poly: &Polynomial) -> usize

Function to calculate the homomorphic length of an encrypted string. This is the number of non-zero coefficients in the encrypted polynomial.

pub fn concatenate_encrypted_strings( &self, encrypted_poly1: &Polynomial, encrypted_poly2: &Polynomial, ) -> Polynomial

Function to concatenate two encrypted strings by combining their encrypted polynomials. This will create a new polynomial containing the coefficients of both encrypted strings.

pub fn extract_encrypted_substring<R>( &self, encrypted_poly: &Polynomial, range: R, ) -> Polynomial
where R: RangeBounds<usize>,

Extracts a substring from an encrypted string, using Rust range syntax.

encrypted_poly: The encrypted string as a polynomial.

range: A range representing the indices to extract (e.g., 0..3, 3.., ..5).